requirements.txt
Flask>=3.0
bcrypt>=4.0
PyYAML>=6.0
cryptography>=42.0
